Cast the refusal code, which the game reports unsigned
GetActionStatus returns uint and the entry carries an int. Caught by the
plugin build, which is the only place it could be: the rotation engine
does not reference ClientStructs.

+++ src/OneTwoPunch.Plugin/Services/ActionStateAdapter.cs
@@ -204,6 +204,6 @@ public sealed unsafe class ActionStateAdapter : IActionState
         const bool unlocked = true;
 
         return new Entry(
-            unlocked, remaining, charges, maxCharges, usable, statusIgnoringRecast == 0, status);
+            unlocked, remaining, charges, maxCharges, usable, statusIgnoringRecast == 0, (int)status);
     }
 }
